some changes for the 2016 data taking
Change the return of all previous method to json (instead of just a string representation). Add a new method to figure out if a run/stream is fully processed (repack or express). Finally add a method to return the list of locked datasets (which is basically a list of datasets produced in the Tier0).